[cleanup] clean separate component logics
